1.1.0

+++ src/client/kotlin/tech/thatgravyboat/skycubed/api/displays/Display.kt
@@ -2,6 +2,7 @@ package tech.thatgravyboat.skycubed.api.displays
 
 import net.minecraft.client.gui.GuiGraphics
 import tech.thatgravyboat.skycubed.utils.pushPop
+import tech.thatgravyboat.skycubed.utils.translate
 
 interface Display {
 
@@ -12,7 +13,7 @@ interface Display {
 
     fun render(graphics: GuiGraphics, x: Int, y: Int, alignmentX: Float = 0f, alignmentY: Float = 0f) {
         graphics.pushPop {
-            translate((x - getWidth() * alignmentX).toDouble(), (y - getHeight() * alignmentY).toDouble(), 0.0)
+            translate((x - getWidth() * alignmentX).toInt(), (y - getHeight() * alignmentY).toInt(), 0)
             render(graphics)
         }
     }
